moved all the interfaces into the models folder. not sure if we are supposed to do this.
